Refactor 'makeLayers'
