+++ javascript/src/make-layers.js
@@ -1,52 +1,8 @@
-import makeDataAccessor from "./helpers/make-data-accessor";
+import LayerBuilder from "./layer-builder";
 
 export default function(el, layers) {
-  return layers.map(function(item) {
-    const properties = item.properties;
-    makeDataAccessors(properties);
-    if (properties.dataframeToD3) {
-      item.data = HTMLWidgets.dataframeToD3(item.data);
-    }
-
-    if (HTMLWidgets.shinyMode) {
-      properties.onClick = function(info) {
-        const data = { lng: info.lngLat[0], lat: info.lngLat[1], object: info.object };
-        Shiny.onInputChange(el.id + "_onclick", data);
-      };
-    }
-
-    if (properties.getTooltip) {
-      const tooltipElement = _deckGLWidget._store[el.id].tooltipElement;
-      setTooltip(properties, tooltipElement);
-    }
-
-    properties.data = item.data;
-    return new deck[item.className](properties);
+  return layers.map(layer => {
+    let layerBuilder = new LayerBuilder(el, layer);
+    return layerBuilder.render();
   });
 }
-
-const makeDataAccessors = function(properties) {
-  for (let key of Object.keys(properties)) {
-    var property = properties[key];
-    if (typeof property === "object" && property.dataAccessor !== undefined) {
-      properties[key] = makeDataAccessor(key, property.dataAccessor);
-    }
-  }
-};
-
-const setTooltip = function(properties, tooltipElement) {
-  properties.onHover = function({ x, y, object }) {
-    if (!object) {
-      tooltipElement.innerHTML = "";
-      return;
-    }
-
-    const text = properties.getTooltip(object);
-    if (!properties.fixedTooltip) {
-      tooltipElement.style.top = y + "px";
-      tooltipElement.style.left = x + "px";
-    }
-
-    tooltipElement.innerHTML = text;
-  };
-};
